separate fastapi app and startup
